refactor(tests): consolidate route and config builders into per-crate test_support modules
Pure test relocation closing the ROADMAP tech-debt item on test fixture
extraction; no production code or test-semantics changes.

crates/rib gains a private #[cfg(test)] test_support module holding the
canonical Route / FlowSpecRoute unit-test constructors (make_route,
make_v6_route, make_route_with_lp, make_route_with_path_id,
make_flowspec_route) that were copied, with drift, across the manager,
Adj-RIB-In, Adj-RIB-Out, and Loc-RIB test modules.

The daemon bin gains src/test_support.rs holding the EVPN primitives
(vni, rd, mac, ip, evpn_instance, gather_metrics_text) duplicated across
five EVPN test modules, and the FIB builders (fib_table, basic_fib_table,
fib_route_key, route_from_peer, scoped_route, route_event) duplicated
across the fib, fib_runtime, blackhole, fib_table_control, and
config_transaction_control test modules.

Where copies had drifted (flowspec peer parameter, EVPN instance ASN
65000 vs 65001, route-event peer address), the canonical form is the
superset and call sites pass their historical values explicitly via thin
local wrappers, so no test asserts anything different.

Deliberately left local: bench/integration-test helpers (separate
compilation units; sharing would require new public surface), the
blackhole community-route builder, the EVPN MAC/IP route family
(divergent rd/sticky/esi axes), and per-module minimal-config TOML
snippets. crates/cli already shares its fixtures via an existing
test_support module; crates/api was done previously.
