+++ crates/rib/src/manager/tests.rs
@@ -5,15 +5,16 @@ use std::time::{Duration, Instant};
 
 use rustbgpd_wire::{
     AddressPrefixOrf, Afi, AsPath, AsPathSegment, EthernetSegmentIdentifier, EthernetTagId,
-    EvpnImet, EvpnMacIp, EvpnRoute, ExtendedCommunity, FlowSpecComponent, FlowSpecPrefix,
-    FlowSpecRule, Ipv4Prefix, Ipv6Prefix, MacAddress, MplsLabel, OrfAction, OrfMatch, Origin,
-    PathAttribute, Prefix, RouteDistinguisher, RpkiValidation, Safi, WhenToRefresh,
+    EvpnImet, EvpnMacIp, EvpnRoute, ExtendedCommunity, Ipv4Prefix, Ipv6Prefix, MacAddress,
+    MplsLabel, OrfAction, OrfMatch, Origin, PathAttribute, Prefix, RouteDistinguisher,
+    RpkiValidation, Safi, WhenToRefresh,
 };
 use tokio::sync::oneshot;
 
 use super::*;
 use crate::event::RouteEventType;
 use crate::route::{EvpnRibRoute, FlowSpecRoute, NextHopScope, Route};
+use crate::test_support::{make_flowspec_route, make_route, make_route_with_lp, make_v6_route};
 
 fn evpn_sendable() -> Vec<(Afi, Safi)> {
     vec![(Afi::L2Vpn, Safi::Evpn)]
@@ -282,98 +283,12 @@ fn assert_refresh_metrics(
     );
 }
 
-fn make_route(prefix: Ipv4Prefix, next_hop: Ipv4Addr) -> Route {
-    Route {
-        prefix: Prefix::V4(prefix),
-        next_hop: IpAddr::V4(next_hop),
-        link_local_next_hop: None,
-        next_hop_scope: None,
-        peer: IpAddr::V4(next_hop),
-        attributes: Arc::new(vec![]),
-        received_at: Instant::now(),
-        origin_type: crate::route::RouteOrigin::Ebgp,
-        peer_router_id: Ipv4Addr::UNSPECIFIED,
-        is_stale: false,
-        is_llgr_stale: false,
-        path_id: 0,
-        validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-        aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-        aspa_context: rustbgpd_wire::AspaValidationContext::default(),
-    }
-}
-
 fn make_indexed_route(index: u32, next_hop: Ipv4Addr) -> Route {
     let octets = index.to_be_bytes();
     make_route(
         Ipv4Prefix::new(Ipv4Addr::new(10, octets[1], octets[2], octets[3]), 32),
         next_hop,
     )
-}
-
-fn make_v6_route(prefix: Ipv6Prefix, next_hop: Ipv6Addr) -> Route {
-    Route {
-        prefix: Prefix::V6(prefix),
-        next_hop: IpAddr::V6(next_hop),
-        link_local_next_hop: None,
-        next_hop_scope: None,
-        peer: IpAddr::V6(next_hop),
-        attributes: Arc::new(vec![]),
-        received_at: Instant::now(),
-        origin_type: crate::route::RouteOrigin::Ebgp,
-        peer_router_id: Ipv4Addr::UNSPECIFIED,
-        is_stale: false,
-        is_llgr_stale: false,
-        path_id: 0,
-        validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-        aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-        aspa_context: rustbgpd_wire::AspaValidationContext::default(),
-    }
-}
-
-fn make_route_with_lp(prefix: Ipv4Prefix, peer: Ipv4Addr, local_pref: u32) -> Route {
-    Route {
-        prefix: Prefix::V4(prefix),
-        next_hop: IpAddr::V4(peer),
-        link_local_next_hop: None,
-        next_hop_scope: None,
-        peer: IpAddr::V4(peer),
-        attributes: Arc::new(vec![
-            PathAttribute::Origin(Origin::Igp),
-            PathAttribute::AsPath(AsPath {
-                segments: vec![AsPathSegment::AsSequence(vec![65001])],
-            }),
-            PathAttribute::LocalPref(local_pref),
-        ]),
-        received_at: Instant::now(),
-        origin_type: crate::route::RouteOrigin::Ebgp,
-        peer_router_id: Ipv4Addr::UNSPECIFIED,
-        is_stale: false,
-        is_llgr_stale: false,
-        path_id: 0,
-        validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-        aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-        aspa_context: rustbgpd_wire::AspaValidationContext::default(),
-    }
-}
-
-fn make_flowspec_route(peer: Ipv4Addr) -> FlowSpecRoute {
-    let prefix = Ipv4Prefix::new(Ipv4Addr::new(192, 0, 2, 0), 24);
-    FlowSpecRoute {
-        rule: FlowSpecRule {
-            components: vec![FlowSpecComponent::DestinationPrefix(FlowSpecPrefix::V4(
-                prefix,
-            ))],
-        },
-        afi: Afi::Ipv4,
-        peer: IpAddr::V4(peer),
-        attributes: vec![],
-        received_at: Instant::now(),
-        origin_type: crate::route::RouteOrigin::Ebgp,
-        peer_router_id: Ipv4Addr::UNSPECIFIED,
-        is_stale: false,
-        is_llgr_stale: false,
-        path_id: 0,
-    }
 }
 
 #[tokio::test]
