+++ crates/rib/src/test_support.rs
@@ -1,0 +1,128 @@
+//! Shared fixtures for this crate's unit tests.
+//!
+//! Canonical test-only route constructors used by the manager,
+//! Adj-RIB-In, Adj-RIB-Out, and Loc-RIB test modules. Keep new
+//! test-only `Route` / `FlowSpecRoute` builders here so a field
+//! addition touches one place instead of one copy per test module.
+//!
+//! Bench and integration-test helpers (`benches/`, `tests/`) compile
+//! as separate units and cannot see this private module; they keep
+//! their own builders.
+
+use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
+use std::sync::Arc;
+use std::time::Instant;
+
+use rustbgpd_wire::{
+    Afi, AsPath, AsPathSegment, FlowSpecComponent, FlowSpecPrefix, FlowSpecRule, Ipv4Prefix,
+    Ipv6Prefix, Origin, PathAttribute, Prefix,
+};
+
+use crate::route::{FlowSpecRoute, Route, RouteOrigin};
+
+pub(crate) fn make_route(prefix: Ipv4Prefix, next_hop: Ipv4Addr) -> Route {
+    Route {
+        prefix: Prefix::V4(prefix),
+        next_hop: IpAddr::V4(next_hop),
+        link_local_next_hop: None,
+        next_hop_scope: None,
+        peer: IpAddr::V4(next_hop),
+        attributes: Arc::new(vec![]),
+        received_at: Instant::now(),
+        origin_type: RouteOrigin::Ebgp,
+        peer_router_id: Ipv4Addr::UNSPECIFIED,
+        is_stale: false,
+        is_llgr_stale: false,
+        path_id: 0,
+        validation_state: rustbgpd_wire::RpkiValidation::NotFound,
+        aspa_state: rustbgpd_wire::AspaValidation::Unknown,
+        aspa_context: rustbgpd_wire::AspaValidationContext::default(),
+    }
+}
+
+pub(crate) fn make_v6_route(prefix: Ipv6Prefix, next_hop: Ipv6Addr) -> Route {
+    Route {
+        prefix: Prefix::V6(prefix),
+        next_hop: IpAddr::V6(next_hop),
+        link_local_next_hop: None,
+        next_hop_scope: None,
+        peer: IpAddr::V6(next_hop),
+        attributes: Arc::new(vec![]),
+        received_at: Instant::now(),
+        origin_type: RouteOrigin::Ebgp,
+        peer_router_id: Ipv4Addr::UNSPECIFIED,
+        is_stale: false,
+        is_llgr_stale: false,
+        path_id: 0,
+        validation_state: rustbgpd_wire::RpkiValidation::NotFound,
+        aspa_state: rustbgpd_wire::AspaValidation::Unknown,
+        aspa_context: rustbgpd_wire::AspaValidationContext::default(),
+    }
+}
+
+pub(crate) fn make_route_with_lp(prefix: Ipv4Prefix, peer: Ipv4Addr, local_pref: u32) -> Route {
+    Route {
+        prefix: Prefix::V4(prefix),
+        next_hop: IpAddr::V4(peer),
+        link_local_next_hop: None,
+        next_hop_scope: None,
+        peer: IpAddr::V4(peer),
+        attributes: Arc::new(vec![
+            PathAttribute::Origin(Origin::Igp),
+            PathAttribute::AsPath(AsPath {
+                segments: vec![AsPathSegment::AsSequence(vec![65001])],
+            }),
+            PathAttribute::LocalPref(local_pref),
+        ]),
+        received_at: Instant::now(),
+        origin_type: RouteOrigin::Ebgp,
+        peer_router_id: Ipv4Addr::UNSPECIFIED,
+        is_stale: false,
+        is_llgr_stale: false,
+        path_id: 0,
+        validation_state: rustbgpd_wire::RpkiValidation::NotFound,
+        aspa_state: rustbgpd_wire::AspaValidation::Unknown,
+        aspa_context: rustbgpd_wire::AspaValidationContext::default(),
+    }
+}
+
+/// Add-path shaped route: fixed next-hop/peer, caller-chosen path id.
+pub(crate) fn make_route_with_path_id(prefix: Prefix, path_id: u32) -> Route {
+    Route {
+        prefix,
+        next_hop: IpAddr::V4(Ipv4Addr::new(10, 0, 0, 1)),
+        link_local_next_hop: None,
+        next_hop_scope: None,
+        peer: IpAddr::V4(Ipv4Addr::new(1, 1, 1, 1)),
+        attributes: Arc::new(vec![]),
+        received_at: Instant::now(),
+        origin_type: RouteOrigin::Ebgp,
+        peer_router_id: Ipv4Addr::new(1, 1, 1, 1),
+        is_stale: false,
+        is_llgr_stale: false,
+        validation_state: rustbgpd_wire::RpkiValidation::NotFound,
+        aspa_state: rustbgpd_wire::AspaValidation::Unknown,
+        aspa_context: rustbgpd_wire::AspaValidationContext::default(),
+        path_id,
+    }
+}
+
+pub(crate) fn make_flowspec_route(peer: Ipv4Addr) -> FlowSpecRoute {
+    let prefix = Ipv4Prefix::new(Ipv4Addr::new(192, 0, 2, 0), 24);
+    FlowSpecRoute {
+        rule: FlowSpecRule {
+            components: vec![FlowSpecComponent::DestinationPrefix(FlowSpecPrefix::V4(
+                prefix,
+            ))],
+        },
+        afi: Afi::Ipv4,
+        peer: IpAddr::V4(peer),
+        attributes: vec![],
+        received_at: Instant::now(),
+        origin_type: RouteOrigin::Ebgp,
+        peer_router_id: Ipv4Addr::UNSPECIFIED,
+        is_stale: false,
+        is_llgr_stale: false,
+        path_id: 0,
+    }
+}
